+++ modules/frontend/combiner/interface.go
@@ -7,9 +7,9 @@ import (
 // Combiner is used to merge multiple responses into a single response.
 //
 // Implementations must be thread-safe.
-// TODO: StatusCode() and the tenant parameter on AddRequest are only used in for multi-tenant support. Can we remove them?
+// TODO: StatusCode() is only used for multi-tenant support. Can we remove it?
 type Combiner interface {
-	AddResponse(r *http.Response, tenant string) error
+	AddResponse(r *http.Response) error
 	StatusCode() int
 	ShouldQuit() bool