+++ modules/frontend/combiner/trace_by_id.go
@@ -2,42 +2,47 @@ package combiner
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
 	"io"
 	"net/http"
 	"strings"
 	"sync"
 
+	"github.com/gogo/protobuf/jsonpb"
 	"github.com/gogo/protobuf/proto"
 	"github.com/grafana/tempo/pkg/api"
 	"github.com/grafana/tempo/pkg/model/trace"
 	"github.com/grafana/tempo/pkg/tempopb"
-	v1 "github.com/grafana/tempo/pkg/tempopb/common/v1"
 )
 
 const (
 	internalErrorMsg = "internal error"
-	tenantLabel      = "tenant"
 )
 
 type traceByIDCombiner struct {
 	mu sync.Mutex
 
-	c *trace.Combiner
+	c           *trace.Combiner
+	contentType string
 
 	code          int
 	statusMessage string
 }
 
-func NewTraceByID() Combiner {
+// NewTraceByID returns a trace id combiner. The trace by id combiner has a few different behaviors then the others
+// - 404 is a valid response code. if all downstream jobs return 404 then it will return 404 with no body
+// - translate tempopb.TraceByIDResponse to tempopb.Trace. all other combiners pass the same object through
+// - runs the zipkin dedupe logic on the fully combined trace
+// - encode the returned trace as either json or proto depending on the request
+func NewTraceByID(maxBytes int, contentType string) Combiner {
 	return &traceByIDCombiner{
-		c:    trace.NewCombiner(0),
-		code: http.StatusNotFound,
+		c:           trace.NewCombiner(maxBytes),
+		code:        http.StatusNotFound,
+		contentType: contentType,
 	}
 }
 
-func (c *traceByIDCombiner) AddResponse(res *http.Response, tenant string) error {
+func (c *traceByIDCombiner) AddResponse(res *http.Response) error {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
@@ -57,7 +62,7 @@ func (c *traceByIDCombiner) AddResponse(res *http.Response, tenant string) error
 			return fmt.Errorf("error reading response body: %w", err)
 		}
 		c.statusMessage = string(bytesMsg)
-		return errors.New(c.statusMessage)
+		return nil
 	}
 
 	// Read the body
@@ -69,18 +74,15 @@ func (c *traceByIDCombiner) AddResponse(res *http.Response, tenant string) error
 	_ = res.Body.Close()
 
 	// Unmarshal the body
-	trace := &tempopb.Trace{}
-	err = trace.Unmarshal(buff)
+	resp := &tempopb.TraceByIDResponse{}
+	err = resp.Unmarshal(buff)
 	if err != nil {
 		c.statusMessage = internalErrorMsg
 		return fmt.Errorf("error unmarshalling response body: %w", err)
 	}
 
-	// inject tenant label as resource in trace
-	InjectTenantResource(tenant, trace)
-
 	// Consume the trace
-	_, err = c.c.Consume(trace)
+	_, err = c.c.Consume(resp.Trace)
 	return err
 }
 
@@ -88,10 +90,10 @@ func (c *traceByIDCombiner) HTTPFinal() (*http.Response, error) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
-	statusCode := c.getStatusCode()
+	statusCode := c.code
 	traceResult, _ := c.c.Result()
 
-	if traceResult == nil || statusCode != http.StatusOK {
+	if statusCode != http.StatusOK {
 		return &http.Response{
 			StatusCode: statusCode,
 			Body:       io.NopCloser(strings.NewReader(c.statusMessage)),
@@ -99,15 +101,36 @@ func (c *traceByIDCombiner) HTTPFinal() (*http.Response, error) {
 		}, nil
 	}
 
-	buff, err := proto.Marshal(traceResult)
+	// if we have no trace result just substitute and return an empty trace
+	if traceResult == nil {
+		traceResult = &tempopb.Trace{}
+	}
+
+	// dedupe duplicate span ids
+	deduper := newDeduper()
+	traceResult = deduper.dedupe(traceResult)
+
+	// marshal in the requested format
+	var buff []byte
+	var err error
+
+	if c.contentType == api.HeaderAcceptProtobuf {
+		buff, err = proto.Marshal(traceResult)
+	} else {
+		var jsonStr string
+
+		marshaler := &jsonpb.Marshaler{}
+		jsonStr, err = marshaler.MarshalToString(traceResult)
+		buff = []byte(jsonStr)
+	}
 	if err != nil {
-		return &http.Response{}, fmt.Errorf("error marshalling response to proto: %w", err)
+		return &http.Response{}, fmt.Errorf("error marshalling response: %w content type: %s", err, c.contentType)
 	}
 
 	return &http.Response{
 		StatusCode: http.StatusOK,
 		Header: http.Header{
-			api.HeaderContentType: {api.HeaderAcceptProtobuf},
+			api.HeaderContentType: {c.contentType},
 		},
 		Body:          io.NopCloser(bytes.NewReader(buff)),
 		ContentLength: int64(len(buff)),
@@ -117,19 +140,7 @@ func (c *traceByIDCombiner) HTTPFinal() (*http.Response, error) {
 func (c *traceByIDCombiner) StatusCode() int {
 	c.mu.Lock()
 	defer c.mu.Unlock()
-	return c.getStatusCode()
-}
-
-func (c *traceByIDCombiner) getStatusCode() int {
-	statusCode := c.code
-	// Translate non-404s 4xx into 500s. If, for instance, we get a 400 back from an internal component
-	// it means that we created a bad request. 400 should not be propagated back to the user b/c
-	// the bad request was due to a bug on our side, so return 500 instead.
-	if statusCode/100 == 4 && statusCode != http.StatusNotFound {
-		statusCode = 500
-	}
-
-	return statusCode
+	return c.code
 }
 
 // ShouldQuit returns true if the response should be returned early.
@@ -140,28 +151,20 @@ func (c *traceByIDCombiner) ShouldQuit() bool {
 }
 
 func (c *traceByIDCombiner) shouldQuit() bool {
-	if c.getStatusCode()/100 == 5 { // Bail on 5xx
+	if c.code/100 == 5 { // Bail on 5xx
+		return true
+	}
+
+	// test special case for 404
+	if c.code == http.StatusNotFound {
+		return false
+	}
+
+	// bail on other 400s
+	if c.code/100 == 4 {
 		return true
 	}
 
 	// 2xx and 404 are OK
 	return false
-}
-
-// InjectTenantResource will add tenantLabel attribute into response to show which tenant the response came from
-func InjectTenantResource(tenant string, t *tempopb.Trace) {
-	if t == nil || t.Batches == nil {
-		return
-	}
-
-	for _, b := range t.Batches {
-		b.Resource.Attributes = append(b.Resource.Attributes, &v1.KeyValue{
-			Key: tenantLabel,
-			Value: &v1.AnyValue{
-				Value: &v1.AnyValue_StringValue{
-					StringValue: tenant,
-				},
-			},
-		})
-	}
 }