+++ modules/frontend/pipeline/collector_grpc.go
@@ -50,7 +50,7 @@ func (c GRPCCollector[T]) RoundTrip(req *http.Request) error {
 		}
 
 		if resp != nil {
-			err = c.combiner.AddResponse(resp, "")
+			err = c.combiner.AddResponse(resp)
 			if err != nil {
 				overallErr = err
 				break